- Fixed some formatting types and variable types for 64-bit compatible version - Fixed a bug with ShaderRequests setting their initial index before finishing the request

+++ ShaderToy/AudioController.mm
@@ -31,7 +31,7 @@ void checkStatus(OSStatus status)
 {
     if (status != 0)
     {
-        printf("Error: %ld\n", status);
+        printf("Error: %d\n", (int)status);
     }
 }
